pyproject.toml
[project]
name = "vts-face-control-plugin"
version = "0.1.0"
description = ""
authors = [
    {name = "Zaxpris",email = "zaxpris@qq.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "websockets (>=15.0.1,<16.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "gradio-client (>=1.8.0,<2.0.0)",
    "pygame (>=2.0.0,<3.0.0)",
    "bilibili-api-python (>=17.2.0,<18.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pycryptodomex (>=3.23.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "brotli (>=1.1.0,<2.0.0)",
    "qrcode[pil] (>=8.2,<9.0)",
    "qrcode-terminal (>=0.8,<0.9)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydub (>=0.25.1,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 128

[tool.ruff]
extend-exclude = []
ignore = [
    "E402",
    "E501",
    "E722",
    "B008",
    "B905",
    "F401",
    "TID252",
    "TRY002",
    "TRY003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF012",
    "RUF100",
    "PGH003",
    "N801",
    "A003",
]
select = [
    "F",
    "E",
    "I",
    "B",
    "A",
    "COM",
    "C4",
    "ISC",
    "PIE",
    "PYI",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "ARG",
    "PTH",
    "PGH",
    "TRY",
    "RUF",
]